[FIX] Correções diversas nas classes
