+++ test/sigep_test/gera_digito_verificador_test.py
@@ -26,14 +26,14 @@
 ###############################################################################
 
 from unittest import TestCase
-from sigep.sigep.gera_digito_verificador import RequestGeraDigitoVerificadorSIGEP
+from sigep.sigep.gera_digito_verificador import \
+    RequestGeraDigitoVerificadorSIGEP
 from sigep.sigep.gera_digito_verificador import ResponseGeraDigitoVerificador
 
 
 class TestRequestGeraDigitoVerificador(TestCase):
 
     def test__init__(self):
-
         etiquetas = 'DL76023727 BR,DL76023728 BR'
         req = RequestGeraDigitoVerificadorSIGEP(etiquetas, 'sigep', 'n5f9t8')
 
@@ -45,7 +45,6 @@ class TestRequestGeraDigitoVerificador(TestCase):
         self.test__init__()
 
     def test_get_xml(self):
-
         etiquetas = 'DL76023727 BR,DL76023728 BR'
         req = RequestGeraDigitoVerificadorSIGEP(etiquetas, 'sigep', 'n5f9t8')
 
@@ -64,7 +63,6 @@ class TestRequestGeraDigitoVerificador(TestCase):
 class TestResponseGeraDigitoVerificador(TestCase):
 
     def test_parse_xml(self):
-
         resp = ResponseGeraDigitoVerificador()
 
         xml = '''<S:Envelope
