+++ test/sigep_test/status_cartao_postagem_test.py
@@ -37,7 +37,8 @@ class TestRequestStatusCartaoPostagem(TestCase):
         senha = 'n5f9t8'
         cartao_postagem = '0057018901'
 
-        req_status = RequestStatusCartaoPostagemSIGEP(cartao_postagem, login, senha)
+        req_status = RequestStatusCartaoPostagemSIGEP(cartao_postagem, login,
+                                                      senha)
 
         xml = req_status.header
         xml += '<cli:getStatusCartaoPostagem>'
